backend/requirements.txt
Django>=3.0,<4.0
psycopg2-binary==2.9.3
django_extra_fields==3.0.2
djangorestframework==3.13.1
drf_extra_fields==3.4.0
setuptools==62.1.0
django-cors-headers==3.13.0
Pillow==9.2.0